completed

+++ controllers/contactController.js
@@ -4,17 +4,17 @@ import Contact from "../models/contactModel.js";
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 export const getContacts = asyncHandler(async(req, res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({user_id: req.user.id})
     res.status(200).json(contacts)
 })
 
 
 //@desc Create a contact
 //@route GET /api/contacts
-//@access public
+//@access private
 
 export const createContact = asyncHandler(async (req, res) => {
     console.log('The request body is: ', req.body);
@@ -25,7 +25,8 @@ export const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id,
     })
 
     res.status(200).json('Create contact success');
@@ -34,7 +35,7 @@ export const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get a contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 export const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
@@ -48,13 +49,19 @@ export const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update a contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 export const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
     if(!contact){
         res.status(404).json('Contact does not exist.')
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error('User not authozirized to update this contact.')
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -66,7 +73,7 @@ export const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete a contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 export const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
@@ -74,6 +81,11 @@ export const deleteContact = asyncHandler(async (req, res) => {
         res.status(400).json('Contact does not exist.')
     }
     
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error('User not authozirized to delete this contact.')
+    }
+
     await Contact.findByIdAndRemove(req.params.id)
     res.status(200).json('Contact deleted.');
 })
